+++ app/src/main/java/com/exemple/constrackerok/Objects/User.java
@@ -5,7 +5,7 @@ import java.io.Serializable;
 public class User implements Serializable {
 
     private static final long serialVersionUID = 1L;
-    private int idUser;
+    private long idUser;
     private String title;
     private String name;
     private String surname;
@@ -17,7 +17,7 @@ public class User implements Serializable {
     public User() {
     }
 
-    public User(int idUser, String title, String name, String surname, String tel, String email, String password) {
+    public User(long idUser, String title, String name, String surname, String tel, String email, String password) {
         this.idUser = idUser;
         this.title = title;
         this.name = name;
@@ -75,11 +75,11 @@ public class User implements Serializable {
         this.name = name;
     }
 
-    public int getIdUser() {
+    public long getIdUser() {
         return idUser;
     }
 
-    public void setIdUser(int idUser) {
+    public void setIdUser(long idUser) {
         this.idUser = idUser;
     }
